refactor: split the alignment code into src/align and test argument validation
Move the alignment implementations under src/align/ (linspace, force_start,
traceback, InteractionAlignment, and alignment.h as symbols.h) and put a
dispatch.h in front of them, so risearch.cpp asks for an alignment instead of
choosing between two implementations itself. RunningMax gives the best-hit
score/position/row a name rather than four parallel locals.

risearch.cpp drops from 232 lines of main() to 80: sequence encoding becomes
encode_sequence(), and the force-start and weights argument checks - which
repeated the same fprintf four times - are single checks now.

The C standard headers become their C++ spellings.

Adds test_exit.cpp covering the four argument combinations that must be
rejected: weights without a force start, a force start without weights, an
extension penalty alongside a force start, and an output mode alongside a
force start. base_args.h holds the shared invocation.

Verified against the previous binary: 44/44 unit tests, and byte-identical
output over 60 flag combinations (5 matrices x default/-p1/-p2/-p3/-s/-d/-n/
-e/-l/-R) on a 3x1200 nt target and 5x21 nt query set.

+++ RIsearch1/src/debug_print.h
@@ -32,7 +32,7 @@
     for (i = 0; i < rows; i++) {
         printf("\n%c", (i == 0 ? '-' : index2nt(*(seq1 + i - 1))));
         for (j = 0; j < cols; j++) {
-            printf("\t%f", (mat[i][j] == NEGINF ? -8 : mat[i][j])); /* -8 as dummy for -inf */
+            printf("\t%f", static_cast<double>(mat[i][j] == NEGINF ? -8 : mat[i][j])); /* -8 as dummy for -inf */
         }
     }
     printf("\n\n");
